Issue #129: Support for server-sent events (SSE)
Original PR #130 by @madeddie

This patch adds a proxy.flushinterval option which
enables periodic flushing of the repsonse buffer.

Since this is a global option and the implications on
non-streaming HTTP requests are unknown the current
implementation limits the use of the flush interval to
requests where the 'Accept' header is set to
'text/event-stream' which identifies SSE requests.

This is really a route specific option and should be
configured as such once this becomes possible.